feat: add parameter guards

+++ user.go
@@ -2,6 +2,7 @@ package passage
 
 import (
 	"context"
+	"errors"
 	"net/http"
 	"strings"
 )
@@ -24,6 +25,10 @@ func newUser(appID string, client *ClientWithResponses) *user {
 
 // Get retrieves a user's object using their user ID.
 func (u *user) Get(userID string) (*PassageUser, error) {
+	if userID == "" {
+		return nil, errors.New("userID is required.")
+	}
+
 	res, err := u.client.GetUserWithResponse(context.Background(), u.appID, userID)
 	if err != nil {
 		return nil, err
@@ -38,6 +43,10 @@ func (u *user) Get(userID string) (*PassageUser, error) {
 
 // GetByIdentifier retrieves a user's object using their user identifier.
 func (u *user) GetByIdentifier(identifier string) (*PassageUser, error) {
+	if identifier == "" {
+		return nil, errors.New("identifier is required.")
+	}
+
 	limit := 1
 	lowerIdentifier := strings.ToLower(identifier)
 	res, err := u.client.ListPaginatedUsersWithResponse(
@@ -72,6 +81,10 @@ func (u *user) GetByIdentifier(identifier string) (*PassageUser, error) {
 
 // Activate activates a user using their user ID.
 func (u *user) Activate(userID string) (*PassageUser, error) {
+	if userID == "" {
+		return nil, errors.New("userID is required.")
+	}
+
 	res, err := u.client.ActivateUserWithResponse(context.Background(), u.appID, userID)
 	if err != nil {
 		return nil, err
@@ -86,6 +99,10 @@ func (u *user) Activate(userID string) (*PassageUser, error) {
 
 // Deactivate deactivates a user using their user ID.
 func (u *user) Deactivate(userID string) (*PassageUser, error) {
+	if userID == "" {
+		return nil, errors.New("userID is required.")
+	}
+
 	res, err := u.client.DeactivateUserWithResponse(context.Background(), u.appID, userID)
 	if err != nil {
 		return nil, err
@@ -100,6 +117,10 @@ func (u *user) Deactivate(userID string) (*PassageUser, error) {
 
 // Update updates a user.
 func (u *user) Update(userID string, options UpdateUserOptions) (*PassageUser, error) {
+	if userID == "" {
+		return nil, errors.New("userID is required.")
+	}
+
 	res, err := u.client.UpdateUserWithResponse(context.Background(), u.appID, userID, options)
 	if err != nil {
 		return nil, err
@@ -114,6 +135,10 @@ func (u *user) Update(userID string, options UpdateUserOptions) (*PassageUser, e
 
 // Create creates a user.
 func (u *user) Create(args CreateUserArgs) (*PassageUser, error) {
+	if args.Email == "" && args.Phone == "" {
+		return nil, errors.New("At least one of args.Email or args.Phone is required.")
+	}
+
 	res, err := u.client.CreateUserWithResponse(context.Background(), u.appID, args)
 	if err != nil {
 		return nil, err
@@ -128,6 +153,10 @@ func (u *user) Create(args CreateUserArgs) (*PassageUser, error) {
 
 // Delete deletes a user using their user ID.
 func (u *user) Delete(userID string) error {
+	if userID == "" {
+		return errors.New("userID is required.")
+	}
+
 	res, err := u.client.DeleteUserWithResponse(context.Background(), u.appID, userID)
 	if err != nil {
 		return err
@@ -142,6 +171,10 @@ func (u *user) Delete(userID string) error {
 
 // ListDevices retrieves a user's webauthn devices using their user ID.
 func (u *user) ListDevices(userID string) ([]WebAuthnDevices, error) {
+	if userID == "" {
+		return nil, errors.New("userID is required.")
+	}
+
 	res, err := u.client.ListUserDevicesWithResponse(context.Background(), u.appID, userID)
 	if err != nil {
 		return nil, err
@@ -156,6 +189,14 @@ func (u *user) ListDevices(userID string) ([]WebAuthnDevices, error) {
 
 // RevokeDevice revokes user's webauthn device using their user ID and the device ID.
 func (u *user) RevokeDevice(userID string, deviceID string) error {
+	if userID == "" {
+		return errors.New("userID is required.")
+	}
+
+	if deviceID == "" {
+		return errors.New("deviceID is required.")
+	}
+
 	res, err := u.client.DeleteUserDevicesWithResponse(context.Background(), u.appID, userID, deviceID)
 	if err != nil {
 		return err
@@ -170,6 +211,10 @@ func (u *user) RevokeDevice(userID string, deviceID string) error {
 
 // RevokeRefreshTokens revokes all of a user's Refresh Tokens using their User ID.
 func (u *user) RevokeRefreshTokens(userID string) error {
+	if userID == "" {
+		return errors.New("userID is required.")
+	}
+
 	res, err := u.client.RevokeUserRefreshTokensWithResponse(context.Background(), u.appID, userID)
 	if err != nil {
 		return err
